% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reaction.R
\name{cpdReactions}
\alias{cpdReactions}
\title{Get all reactions with a specific compound}
\usage{
cpdReactions(db, cpd)
}
\arguments{
\item{db}{Database of class \link(gapseqDB)}

\item{cpd}{Compound ID}
}
\value{
A list with four character vectors (consuming, producing,
consuming+producing, transport)
}
\description{
Find all reaction where a specific compound participates grouped
by consuming, producing, consuming+producing, transport.
}
