% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reaction.R
\name{findReactions}
\alias{findReactions}
\title{Find reaction(s)}
\usage{
findReactions(
  db,
  cpd = NULL,
  cpd.link = "AND",
  ec = NULL,
  ec.link = "AND",
  grep.str = NULL,
  grep.str.link = "AND",
  global.link = "AND"
)
}
\arguments{
\item{db}{Database of class \link(gapseqDB)}

\item{cpd}{Compound IDs to look for.}

\item{cpd.link}{"AND" or "OR". "AND" requires that all IDs in 'cpd.link'
occur in a reaction. "OR" requires only one compound to be present.}
}
\description{
Find a reaction by single or combined criteria.
}
