% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reaction.R
\name{findRxns}
\alias{findRxns}
\title{Find reaction(s)}
\usage{
findRxns(
  db,
  cpd = NULL,
  cpd.link = "AND",
  ec = NULL,
  grep.str = NULL,
  global.link = "AND"
)
}
\arguments{
\item{db}{Database of class \link{gapseqDB-class}}

\item{cpd}{Compound IDs to look for.}

\item{cpd.link}{"AND" or "OR". "AND" requires that all IDs in 'cpd.link'
occur in a reaction. "OR" requires only one compound to be present.}

\item{ec}{EC number to look for.}

\item{grep.str}{Regular expression applied to the reaction equation.}
}
\description{
Find a reaction by single or combined criteria.
}
