% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapseqDB.R
\docType{class}
\name{gapseqDB-class}
\alias{gapseqDB-class}
\title{S4 class for gapseq's biochemistry database}
\description{
S4 class for gapseq's biochemistry database
}
\section{Slots}{

\describe{
\item{\code{rxn}}{A list of tables (\link(data.table)) with reaction information}

\item{\code{cpd}}{A list of tables (\link(data.table)) with compound information}
}}

