% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reaction.R
\name{getChargeBalance}
\alias{getChargeBalance}
\title{Calculate charge balance}
\usage{
getChargeBalance(db, rxn)
}
\arguments{
\item{db}{Database of class \link(gapseqDB)}

\item{rxn}{Reaction ID(s)}
}
\description{
Calculates the charge balance for reaction(s)
}
