% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reaction.R
\name{getMassBalance}
\alias{getMassBalance}
\title{Calculate mass balance}
\usage{
getMassBalance(db, rxn)
}
\arguments{
\item{db}{Database of class \link{gapseqDB-class}}

\item{rxn}{Reaction ID(s)}
}
\description{
Calculates the mass balance for reaction(s)
}
