% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapseqDB.R
\name{initDB}
\alias{initDB}
\title{Initialize a gapseq biochemistry database object}
\usage{
initDB(src = NULL)
}
\arguments{
\item{src}{Location of gagpseq (see details)}
}
\description{
Biochemistry data is retrieved either from gapseq's main github
repository or from a local directory
}
\details{
If src is NULL, the latest database files are retrieved from
the latest commit on the master branch of gapseq from the github repository
(https://github.com/jotech/gapseq).

If 'src' is a git commit SHA fingerprint, the database files are retrieved
from the master branch from github for the specified commit.

If 'src' is a path to a local gapseq installation directory, database
files are directly retrieved from there.
}
