% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,gapseqDB-method}
\alias{summary,gapseqDB-method}
\title{Summary for a compound/reaction from gapseq biochemistry database}
\usage{
\S4method{summary}{gapseqDB}(object, id)
}
\arguments{
\item{object}{Database of class \link(gapseqDB)}

\item{id}{ID of a compound or reaction}
}
\description{
Prints a summary for a reaction or compound to console prompt
}
\details{
Compound IDs should have the prefix "cpd" and reactions "rxn"
}
